feat: HasTicketsSoldOutException e ticketClient para validação

+++ ms-event-manager/src/main/java/io/github/juandev/mseventmanager/service/EventService.java
@@ -1,6 +1,10 @@
 package io.github.juandev.mseventmanager.service;
 
+import feign.FeignException;
+import io.github.juandev.mseventmanager.client.TicketClient;
 import io.github.juandev.mseventmanager.client.ViacepClient;
+import io.github.juandev.mseventmanager.exception.BadRequestException;
+import io.github.juandev.mseventmanager.exception.HasTicketsSoldOutException;
 import io.github.juandev.mseventmanager.exception.NotFoundException;
 import io.github.juandev.mseventmanager.model.Address;
 import io.github.juandev.mseventmanager.model.Event;
@@ -11,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.aggregation.ArithmeticOperators;
 import org.springframework.stereotype.Service;
 
+import java.time.format.DateTimeParseException;
 import java.util.List;
 
 @Service
@@ -19,12 +24,14 @@ public class EventService {
     private final EventRepository repository;
     private final EventMapper mapper;
     private final ViacepClient viacepClient;
+    private final TicketClient ticketClient;
 
     @Autowired
-    public EventService(EventRepository repository, EventMapper mapper, ViacepClient viacepClient) {
+    public EventService(EventRepository repository, EventMapper mapper, ViacepClient viacepClient, TicketClient ticketClient) {
         this.repository = repository;
         this.mapper = mapper;
         this.viacepClient = viacepClient;
+        this.ticketClient = ticketClient;
     }
 
     public Event findById(String id) {
@@ -38,24 +45,30 @@ public class EventService {
     }
 
     public Event save(EventDto eventDto) {
+        try {
+            Event event = mapper.DtoToEvent(eventDto);
 
-        //Fazer uma excessão para cep inválido
-        Event event = mapper.DtoToEvent(eventDto);
+            String cep = event.getCep().replace("-", "");
 
-        String cep = event.getCep().replace("-","");
+            Address address = viacepClient.getAddressByCep(cep);
+            event.setLogradouro(address.getLogradouro());
+            event.setLocalidade(address.getLocalidade());
+            event.setUf(address.getUf());
+            event.setBairro(address.getBairro());
 
-        Address address = viacepClient.getAddressByCep(cep);
-        event.setLogradouro(address.getLogradouro());
-        event.setLocalidade(address.getLocalidade());
-        event.setUf(address.getUf());
-        event.setBairro(address.getBairro());
-
-        return repository.save(event);
+            return repository.save(event);
+        }catch (FeignException.FeignClientException e){
+            throw new BadRequestException("CEP inválido");
+        }
+        //Configurar uma exception para a data
     }
 
     public void deleteById(String id) {
         if (!repository.existsById(id)) {
             throw new NotFoundException("Event with id: "+id+" not found");
+        }
+        if (!ticketClient.checkTicketsByEvent(id).isEmpty()){
+            throw new HasTicketsSoldOutException("Há tickets vendidos, não é possível excluir o evento");
         }
         repository.deleteById(id);
     }
@@ -64,20 +77,23 @@ public class EventService {
         if (!repository.existsById(id)) {
             throw new NotFoundException("Event with id: "+id+" not found");
         }
+        try {
+            Event oldEvent = findById(id);
+            oldEvent.setEventName(event.getEventName());
+            oldEvent.setDateTime(event.getDateTime());
+            oldEvent.setCep(event.getCep());
 
-        Event oldEvent = findById(id);
-        oldEvent.setEventName(event.getEventName());
-        oldEvent.setDateTime(event.getDateTime());
-        oldEvent.setCep(event.getCep());
+            String cep = event.getCep().replace("-", "");
 
-        String cep = event.getCep().replace("-","");
+            Address newAddress = viacepClient.getAddressByCep(cep);
+            oldEvent.setLogradouro(newAddress.getLogradouro());
+            oldEvent.setLocalidade(newAddress.getLocalidade());
+            oldEvent.setUf(newAddress.getUf());
+            oldEvent.setBairro(newAddress.getBairro());
 
-        Address newAddress = viacepClient.getAddressByCep(cep);
-        oldEvent.setLogradouro(newAddress.getLogradouro());
-        oldEvent.setLocalidade(newAddress.getLocalidade());
-        oldEvent.setUf(newAddress.getUf());
-        oldEvent.setBairro(newAddress.getBairro());
-
-        return repository.save(oldEvent);
+            return repository.save(oldEvent);
+        }catch (FeignException.FeignClientException e){
+            throw new BadRequestException("CEP inválido");
+        }
     }
 }
